Changes galore!

+++ aruw-mcb-project/test/aruwsrc/mock/x_drive_chassis_subsystem_mock.hpp
@@ -42,10 +42,10 @@ public:
     MOCK_METHOD(float, chassisSpeedRotationPID, (float, float));
     MOCK_METHOD(void, refresh, (), (override));
     MOCK_METHOD(float, calculateRotationTranslationalGain, (float), ());
-    MOCK_METHOD(int16_t, getLeftFrontRpmActual, (), (const override));
-    MOCK_METHOD(int16_t, getLeftBackRpmActual, (), (const override));
-    MOCK_METHOD(int16_t, getRightFrontRpmActual, (), (const override));
-    MOCK_METHOD(int16_t, getRightBackRpmActual, (), (const override));
+    MOCK_METHOD(float, getLeftFrontRpmActual, (), (const override));
+    MOCK_METHOD(float, getLeftBackRpmActual, (), (const override));
+    MOCK_METHOD(float, getRightFrontRpmActual, (), (const override));
+    MOCK_METHOD(float, getRightBackRpmActual, (), (const override));
     MOCK_METHOD(float, getDesiredRotation, (), (const override));
 };  // class XDriveChassisSubsystemMock
 }  // namespace mock
